app/youtube_service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
youtube-search
pydantic
pydantic_settings>=2.0
python-jose[cryptography]
sqlalchemy
psycopg2-binary
boto3
